+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
 - Added `mw wiki page list` command
 - Added `--dry-run` option to all `mw wiki page` commands
 - Added ability to pass a list of title to `mw wiki page delete`
+- Fixed issue with global command flags being parsed and used in the `docker` command.
 
 ## v0.27.2